+++ packages/components/slideshow/src/translations/loader.ts
@@ -1,4 +1,3 @@
-import type { Translation } from '@288-toolkit/i18n/types';
 import { key } from './index.js';
 
 export const slideshow = {
@@ -7,4 +6,4 @@ export const slideshow = {
 		en: () => import('./en.js'),
 		fr: () => import('./fr.js')
 	}
-} as const satisfies Translation;
+} as const;