+++ packages/forms/src/translations/index.ts
@@ -1,4 +1,3 @@
-import type { Translation } from '@288-toolkit/i18n/types';
 import { key as key_newsletter } from './newsletter-form/index.js';
 
 export const newsletterForm = {
@@ -7,4 +6,4 @@ export const newsletterForm = {
 		en: () => import('./newsletter-form/en.js'),
 		fr: () => import('./newsletter-form/fr.js')
 	}
-} as const satisfies Translation;
+} as const;