+++ packages/components/pagination/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @288-toolkit/pagination
 
+## 3.2.2
+
+### Patch Changes
+
+- Improve exported types
+- Updated dependencies
+  - @288-toolkit/forms@4.3.2
+  - @288-toolkit/i18n@9.0.1
+
 ## 3.2.1
 
 ### Patch Changes